+++ src/product/review/product-review.controller.ts
@@ -1,12 +1,9 @@
 import {
   Body,
   Controller,
-  Delete,
   Get,
-  HttpCode,
   Param,
   ParseIntPipe,
-  Patch,
   Post,
   Query,
   Req,
@@ -15,14 +12,13 @@ import {
 import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 import { plainToInstance } from 'class-transformer';
 
+import { ReviewDto } from '../../review/dto/review.dto';
 import { AccessTokenGuard } from '@/auth/guard/access-token.guard';
 
 import { CreateReviewDto } from './dto/create-review.dto';
 import { FindReviewsDto } from './dto/find-reviews.dto';
 import { ReviewCountDto } from './dto/review-count.dto';
-import { ReviewDto } from './dto/review.dto';
 import { ReviewsDto } from './dto/reviews.dto';
-import { UpdateReviewDto } from './dto/update-review.dto';
 import { ProductReviewService } from './product-review.service';
 
 import { TokenGuardReq } from '@/types/refreshTokenGuard.type';
@@ -64,36 +60,6 @@ export class ProductReviewController {
   }
 
   @ApiOperation({
-    summary: 'Delete a product review'
-  })
-  @ApiResponse({
-    status: 204,
-    description: 'No Content'
-  })
-  @ApiResponse({
-    status: 401,
-    description: 'Unauthorized - Invalid or expired access token, or access token is missing'
-  })
-  @ApiResponse({
-    status: 403,
-    description: 'Forbidden - You are not authorized to delete this review'
-  })
-  @ApiResponse({
-    status: 404,
-    description: 'Not Found - Review not found'
-  })
-  @UseGuards(AccessTokenGuard)
-  @HttpCode(204)
-  @Delete(':id')
-  async remove(
-    @Req() { id }: TokenGuardReq,
-    @Param('productId', ParseIntPipe) productId: number,
-    @Param('id', ParseIntPipe) reviewId: number
-  ) {
-    await this.service.remove(id, productId, reviewId);
-  }
-
-  @ApiOperation({
     summary: 'Get product review count'
   })
   @ApiResponse({
@@ -109,37 +75,6 @@ export class ProductReviewController {
   async count(@Param('productId', ParseIntPipe) productId: number) {
     const count = await this.service.count(productId);
     return plainToInstance(ReviewCountDto, count);
-  }
-
-  @ApiOperation({
-    summary: 'Get a product review'
-  })
-  @ApiResponse({
-    status: 200,
-    description: 'Success',
-    type: ReviewDto
-  })
-  @ApiResponse({
-    status: 401,
-    description: 'Unauthorized - Invalid or expired access token, or access token is missing'
-  })
-  @ApiResponse({
-    status: 403,
-    description: 'Forbidden - You are not authorized to delete this review'
-  })
-  @ApiResponse({
-    status: 404,
-    description: 'Not Found - Review not found'
-  })
-  @UseGuards(AccessTokenGuard)
-  @Get(':id')
-  async findOne(
-    @Req() req: TokenGuardReq,
-    @Param('productId', ParseIntPipe) productId: number,
-    @Param('id', ParseIntPipe) reviewId: number
-  ) {
-    const review = await this.service.findOne(req.id, productId, reviewId);
-    return plainToInstance(ReviewDto, review);
   }
 
   @ApiOperation({
@@ -165,46 +100,5 @@ export class ProductReviewController {
   ) {
     const reviews = await this.service.findMany(findReviewsDto, productId);
     return plainToInstance(ReviewsDto, reviews);
-  }
-
-  @ApiOperation({
-    summary: 'Update a product review'
-  })
-  @ApiResponse({
-    status: 200,
-    description: 'Ok',
-    type: ReviewDto
-  })
-  @ApiResponse({
-    status: 400,
-    description: 'Bad Request - Invalid or missing fields in the request body'
-  })
-  @ApiResponse({
-    status: 401,
-    description: 'Unauthorized - Invalid or expired access token, or access token is missing'
-  })
-  @ApiResponse({
-    status: 403,
-    description: 'Forbidden - You are not authorized to patch this review'
-  })
-  @ApiResponse({
-    status: 404,
-    description: 'Not Found - Review not found'
-  })
-  @UseGuards(AccessTokenGuard)
-  @Patch(':id')
-  async update(
-    @Req() req: TokenGuardReq,
-    @Body() updateReviewDto: UpdateReviewDto,
-    @Param('productId', ParseIntPipe) productId: number,
-    @Param('id', ParseIntPipe) reviewId: number
-  ) {
-    const review = await this.service.update({
-      userId: req.id,
-      productId,
-      id: reviewId,
-      updateReviewDto
-    });
-    return plainToInstance(ReviewDto, review);
   }
 }
