+++ src/product/review/product-review.service.spec.ts
@@ -1,4 +1,4 @@
-import { ConflictException, ForbiddenException, NotFoundException } from '@nestjs/common';
+import { ConflictException, NotFoundException } from '@nestjs/common';
 import { Test } from '@nestjs/testing';
 import { Product } from '@prisma/client';
 import { DeepMockProxy, mockDeep } from 'jest-mock-extended';
@@ -10,7 +10,6 @@ import { getMockProduct, getMockReview } from '@/utils/unit-test';
 
 import { CreateReviewDto } from './dto/create-review.dto';
 import { FindReviewsDto } from './dto/find-reviews.dto';
-import { UpdateReviewDto } from './dto/update-review.dto';
 import { ProductReviewService } from './product-review.service';
 
 import { OrderBy, ReviewsData, SortBy } from '@/types/review.type';
@@ -62,51 +61,6 @@ describe('ProductReviewService', () => {
       return expect(
         service.create(createReviewDto, review.userId, review.productId)
       ).rejects.toThrow(ConflictException);
-    });
-  });
-
-  describe('remove', () => {
-    it('should call prisma delete method when the review exists', async () => {
-      const review = getMockReview();
-      prismaService.review.findUnique.mockResolvedValue(review);
-      prismaService.review.delete.mockResolvedValue(review);
-      await service.remove(review.userId, review.productId, review.id);
-      expect(prismaService.review.delete).toHaveBeenCalled();
-    });
-
-    it('should throw NotFoundException when review is not found', async () => {
-      prismaService.review.findUnique.mockResolvedValue(null);
-      return expect(service.remove('1234567890', 1, 1)).rejects.toThrow(NotFoundException);
-    });
-
-    it('should throw ForbiddenException when user is not authorized to delete the review', async () => {
-      const review = getMockReview();
-      prismaService.review.findUnique.mockResolvedValue(review);
-      return expect(service.remove('0987654321', review.productId, review.id)).rejects.toThrow(
-        ForbiddenException
-      );
-    });
-  });
-
-  describe('findOne', () => {
-    it('should return a review', async () => {
-      const review = getMockReview();
-      prismaService.review.findUnique.mockResolvedValue(review);
-      const result = await service.findOne(review.userId, review.productId, review.id);
-      expect(result).toEqual(review);
-    });
-
-    it('should throw NotFoundException when review is not found', async () => {
-      prismaService.review.findUnique.mockResolvedValue(null);
-      return expect(service.findOne('1234567890', 1, 1)).rejects.toThrow(NotFoundException);
-    });
-
-    it('should throw ForbiddenException when user is not authorized to delete the review', async () => {
-      const review = getMockReview();
-      prismaService.review.findUnique.mockResolvedValue(review);
-      return expect(service.findOne('another-user', review.productId, review.id)).rejects.toThrow(
-        ForbiddenException
-      );
     });
   });
 
@@ -233,44 +187,6 @@ describe('ProductReviewService', () => {
     it('should throw NotFoundException when product is not found', async () => {
       prismaService.product.findUnique.mockResolvedValue(null);
       return expect(service.count(1)).rejects.toThrow(NotFoundException);
-    });
-  });
-
-  describe('update', () => {
-    it('should return a review', async () => {
-      const review = getMockReview();
-      const updateReviewDto: UpdateReviewDto = { text: review.text, rating: review.rating };
-      prismaService.review.findUnique.mockResolvedValue(review);
-      prismaService.review.update.mockResolvedValue(review);
-      const result = await service.update({
-        userId: review.userId,
-        updateReviewDto,
-        productId: review.productId,
-        id: review.id
-      });
-      expect(result).toEqual(review);
-    });
-
-    it('should throw NotFoundException when review is not found', async () => {
-      const updateReviewDto: UpdateReviewDto = { text: 'new-text', rating: 5 };
-      prismaService.review.findUnique.mockResolvedValue(null);
-      return expect(
-        service.update({ userId: '1234567890', updateReviewDto, productId: 1, id: 1 })
-      ).rejects.toThrow(NotFoundException);
-    });
-
-    it('should throw ForbiddenException when user is not authorized to delete the review', async () => {
-      const review = getMockReview();
-      const updateReviewDto: UpdateReviewDto = { text: 'new-text', rating: 5 };
-      prismaService.review.findUnique.mockResolvedValue(review);
-      return expect(
-        service.update({
-          userId: 'another-user',
-          updateReviewDto,
-          productId: review.productId,
-          id: review.id
-        })
-      ).rejects.toThrow(ForbiddenException);
     });
   });
 });
