+++ test/controller/review.integration-spec.ts
@@ -1,12 +1,16 @@
 import { INestApplication } from '@nestjs/common';
+import { Product } from '@prisma/client';
 import * as request from 'supertest';
 
 import { AuthModule } from '@/auth/auth.module';
 import { PrismaService } from '@/prisma/prisma.service';
 import { ReviewsDto } from '@/product/review/dto/reviews.dto';
+import { ReviewDto } from '@/review/dto/review.dto';
+import { UpdateReviewDto } from '@/review/dto/update-review.dto';
 import { ReviewModule } from '@/review/review.module';
 import {
   createProduct,
+  createReview,
   createTestingApp,
   createUser,
   deleteProduct,
@@ -18,28 +22,69 @@ describe('ProductReviewController (integration)', () => {
   let app: INestApplication;
   let prismaService: PrismaService;
 
-  beforeEach(async () => {
+  const userId = '2';
+  const anotherUserId = '3';
+  let accessTokenCookie: string;
+  let anotherAccessTokenCookie: string;
+  let product: Product;
+
+  beforeAll(async () => {
     app = await createTestingApp([ReviewModule, AuthModule]);
     prismaService = app.get(PrismaService);
+
+    const user = await createUser(app, userId);
+    const anotherUser = await createUser(app, anotherUserId);
+    product = await createProduct(prismaService);
+    accessTokenCookie = user.accessTokenCookie;
+    anotherAccessTokenCookie = anotherUser.accessTokenCookie;
   });
 
-  const createReview = async (userId: string, productId: number, rating?: number) => {
-    return prismaService.review.create({
-      data: { userId, productId, text: '', rating: rating || 1 }
-    });
-  };
+  afterAll(async () => {
+    await deleteUser(prismaService, userId);
+    await deleteUser(prismaService, anotherUserId);
+    await deleteProduct(prismaService, product.id);
+  });
 
-  describe('/api/products/:productId/reviews (GET)', () => {
+  describe('/api/reviews/:id (GET)', () => {
     it('200', async () => {
-      const userId1 = '2';
-      const userId2 = '3';
-      const { accessTokenCookie } = await createUser(app, userId1);
-      await createUser(app, userId2);
-      const product = await createProduct(prismaService);
-      await Promise.all([
-        createReview(userId1, product.id, 1),
-        createReview(userId2, product.id, 3)
-      ]);
+      const { id } = await createReview(prismaService, { userId, productId: product.id });
+
+      const { body } = await request(app.getHttpServer())
+        .get(`/api/reviews/${id}`)
+        .set('Cookie', [accessTokenCookie])
+        .expect(200);
+      validateDto(ReviewDto, body);
+
+      await prismaService.review.delete({ where: { id } });
+    });
+
+    it('401', async () => {
+      return request(app.getHttpServer()).get('/api/reviews/0').expect(401);
+    });
+
+    it('403', async () => {
+      const { id } = await createReview(prismaService, { userId, productId: product.id });
+
+      await request(app.getHttpServer())
+        .get(`/api/reviews/${id}`)
+        .set('Cookie', [anotherAccessTokenCookie])
+        .expect(403);
+
+      await prismaService.review.delete({ where: { id } });
+    });
+
+    it('404', async () => {
+      await request(app.getHttpServer())
+        .get('/api/reviews/0')
+        .set('Cookie', [accessTokenCookie])
+        .expect(404);
+    });
+  });
+
+  describe('/api/reviews (GET)', () => {
+    it('200', async () => {
+      const { id } = await createReview(prismaService, { userId, productId: product.id });
+
       const page = 1;
       const { body } = await request(app.getHttpServer())
         .get(`/api/reviews?page=${page}`)
@@ -47,27 +92,114 @@ describe('ProductReviewController (integration)', () => {
         .expect(200);
       validateDto(ReviewsDto, body);
 
-      await Promise.allSettled([
-        deleteUser(prismaService, userId1),
-        deleteUser(prismaService, userId2),
-        deleteProduct(prismaService, product.id)
-      ]);
+      await prismaService.review.delete({ where: { id } });
     });
 
     it('400', async () => {
-      const userId = '2';
-      const { accessTokenCookie } = await createUser(app, userId);
-
       await request(app.getHttpServer())
         .get('/api/reviews?page=a')
         .set('Cookie', [accessTokenCookie])
         .expect(400);
-
-      await deleteUser(prismaService, userId);
     });
 
     it('401', async () => {
       await request(app.getHttpServer()).get('/api/reviews?page=1').expect(401);
+    });
+  });
+
+  describe('/api/reviews/:id (PATCH)', () => {
+    it('200', async () => {
+      const { id } = await createReview(prismaService, { userId, productId: product.id });
+
+      const patchReviewDto: UpdateReviewDto = {
+        text: 'new-text',
+        rating: 5
+      };
+      const { body } = await request(app.getHttpServer())
+        .patch(`/api/reviews/${id}`)
+        .set('Cookie', [accessTokenCookie])
+        .send(patchReviewDto)
+        .expect(200);
+      validateDto(ReviewDto, body);
+
+      await prismaService.review.delete({ where: { id } });
+    });
+
+    it('400', async () => {
+      await request(app.getHttpServer())
+        .patch('/api/reviews/1')
+        .send({
+          text: 'new-text',
+          rating: 0
+        })
+        .set('Cookie', [accessTokenCookie])
+        .expect(400);
+    });
+
+    it('401', async () => {
+      return request(app.getHttpServer()).patch('/api/reviews/0').expect(401);
+    });
+
+    it('403', async () => {
+      const { id } = await createReview(prismaService, { userId, productId: product.id });
+
+      const patchReviewDto: UpdateReviewDto = {
+        text: 'new-text',
+        rating: 5
+      };
+      await request(app.getHttpServer())
+        .patch(`/api/reviews/${id}`)
+        .send(patchReviewDto)
+        .set('Cookie', [anotherAccessTokenCookie])
+        .expect(403);
+
+      await prismaService.review.delete({ where: { id } });
+    });
+
+    it('404', async () => {
+      const patchReviewDto: UpdateReviewDto = {
+        text: 'new-text',
+        rating: 5
+      };
+
+      await request(app.getHttpServer())
+        .patch('/api/reviews/0')
+        .send(patchReviewDto)
+        .set('Cookie', [accessTokenCookie])
+        .expect(404);
+    });
+  });
+
+  describe('/api/reviews/:id (DELETE)', () => {
+    it('204', async () => {
+      const { id } = await createReview(prismaService, { userId, productId: product.id });
+
+      await request(app.getHttpServer())
+        .delete(`/api/reviews/${id}`)
+        .set('Cookie', [accessTokenCookie])
+        .expect(204);
+    });
+
+    it('401', async () => {
+      return request(app.getHttpServer()).delete('/api/reviews/0').expect(401);
+    });
+
+    it('403', async () => {
+      const { id } = await createReview(prismaService, { userId, productId: product.id });
+
+      await request(app.getHttpServer())
+        .delete(`/api/reviews/${id}`)
+        .set('Cookie', [anotherAccessTokenCookie])
+        .expect(403);
+
+      await prismaService.review.delete({ where: { id } });
+    });
+
+    it('404', async () => {
+      await request(app.getHttpServer())
+        .delete('/api/reviews/0')
+        .set('Cookie', [accessTokenCookie])
+        .expect(404);
     });
   });
 });
